requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
ansi2html==1.8.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arize-phoenix==1.2.1
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4>=4.12.3
bleach==6.0.0
boto3==1.29.6
botocore==1.32.6
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.36
dash==2.13.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.3
ddsketch==2.0.4
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
display==1.0.0
distro==1.8.0
docutils==0.20.1
environs==9.5.0
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.2.5
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gcsfs==2023.10.0
google-api-core==2.14.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
graphql-core==3.2.3
greenlet==3.0.1
grpcio==1.56.0
h11==0.14.0
hdbscan==0.8.33
httpcore==1.0.2
httpx==0.25.2
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keyring==24.2.0
kiwisolver==1.4.5
langchain==0.0.341
langchain-core==0.0.6
langsmith==0.0.66
llama-index==0.9.8.post1
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplot==0.1.9
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
minio==7.2.0
mistune==3.0.1
more-itertools==10.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
nh3==0.2.14
nltk==3.8.1
nodeenv==1.8.0
notebook==7.0.3
notebook_shim==0.2.3
numba==0.58.1
numpy==1.25.2
oauthlib==3.2.2
openai==1.3.5
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.10.0
plotly==5.16.1
prisma==0.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.16.1
pyloco==0.0.139
pymilvus==2.2.15
pynndescent==0.5.11
pyparsing==3.0.9
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
readme-renderer==42.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.10.2
rsa==4.9
s3transfer==0.7.0
scikit-learn==1.2.2
scipy==1.11.2
selenium==4.16.0
Send2Trash==1.8.2
SimpleWebSocketServer==0.1.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.2
starlette==0.32.0.post1
strawberry-graphql==0.208.2
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.5.1
tinycss2==1.2.1
tomlkit==0.12.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
trio==0.23.2
trio-websocket==0.11.1
twine==4.0.2
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
umap-learn==0.5.5
uri-template==1.3.0
urllib3==2.0.4
ushlex==0.99.1
uvicorn==0.24.0.post1
wbdata==0.3.0
wbgapi==1.0.12
wcwidth==0.2.6
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.2.3
widgetsnbextension==4.0.8
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.3
zipp==3.16.2

icecream~=2.1.3
schedule~=1.2.1
pytest~=7.4.4
